reasoning-graph-main/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
neo4j==5.15.0
python-dotenv==1.0.0
openai==1.45.0
httpx>=0.23.0,<0.28.0
waitress==3.0.2